[ICU-535] use va_arg(arg, double) and cast down to float or gcc 2.96 will make an error about va_arg(arg, float) since float is promoted to double in ... X-Commit-URL: https://ssl.icu-project.org/trac/changeset/2726

+++ icu4c/source/extra/ustdio/uprintf.c
@@ -1694,7 +1694,7 @@ u_vfprintf_u(    UFILE        *f,
         break;
     
       case ufmt_float:
-        args[cur_arg].floatValue = va_arg(ap, float);
+        args[cur_arg].floatValue = (float) va_arg(ap, double);
         break;
     
       case ufmt_double:
